nuscenesdataset.py now handles some errors (missing files etc). Changed misleading figure name

+++ nuscenesdataset.py
@@ -2,6 +2,7 @@ import os.path as osp
 import os
 import shutil
 import torch
+from pathlib import Path
 
 from torch_geometric.datasets import ModelNet
 from torch_geometric.data import DataLoader, Dataset, extract_zip, Data
@@ -56,76 +57,79 @@ class NuScenesLoader(Dataset):
         self.data, self.slices = torch.load(path)
 
     def get_data(self, idx):
-        this_annotation = self.dataset.sample_annotation[idx]
-        sample_data_tokens = self.dataset.get('sample', this_annotation['sample_token'])['data']
-        sensors = ['RADAR_FRONT', 'RADAR_FRONT_LEFT', 'RADAR_FRONT_RIGHT', 'RADAR_BACK_LEFT', 'RADAR_BACK_RIGHT',
-                   'LIDAR_TOP']
+        try:
+            this_annotation = self.dataset.sample_annotation[idx]
+            sample_data_tokens = self.dataset.get('sample', this_annotation['sample_token'])['data']
+            sensors = ['RADAR_FRONT', 'RADAR_FRONT_LEFT', 'RADAR_FRONT_RIGHT', 'RADAR_BACK_LEFT', 'RADAR_BACK_RIGHT',
+                       'LIDAR_TOP']
 
-        # FIELDS x y z is_lidar intensity vx vy
-        points_ego_frame = [[], [], [], [], [], [], []]
-        for sensor_name in sensors:
-            this_sample_data = self.dataset.get('sample_data', sample_data_tokens[sensor_name])
-            this_calibrated_sensor = self.dataset.get('calibrated_sensor', this_sample_data['calibrated_sensor_token'])
+            # FIELDS x y z is_lidar intensity vx vy
+            points_ego_frame = [[], [], [], [], [], [], []]
+            for sensor_name in sensors:
+                this_sample_data = self.dataset.get('sample_data', sample_data_tokens[sensor_name])
+                this_calibrated_sensor = self.dataset.get('calibrated_sensor', this_sample_data['calibrated_sensor_token'])
 
-            # Get box (sensor coordinate frame)
-            _, boxes, _ = self.dataset.get_sample_data(sample_data_token=sample_data_tokens[sensor_name],
-                                                       selected_anntokens=[this_annotation['token']])
-            box = boxes[0]
+                # Get box (sensor coordinate frame)
+                _, boxes, _ = self.dataset.get_sample_data(sample_data_token=sample_data_tokens[sensor_name],
+                                                           selected_anntokens=[this_annotation['token']])
+                box = boxes[0]
 
-            # Transform box from sensor coordinate frame to ego pose frame
-            box.rotate(Quaternion(this_calibrated_sensor['rotation']))
-            box.translate(np.array(this_calibrated_sensor['translation']))
+                # Transform box from sensor coordinate frame to ego pose frame
+                box.rotate(Quaternion(this_calibrated_sensor['rotation']))
+                box.translate(np.array(this_calibrated_sensor['translation']))
 
-            # Get points (sensor coordinate frame)
-            filename = this_sample_data['filename']
-            if this_sample_data['sensor_modality'] == 'radar':
-                pointcloud = RadarPointCloud.from_file(osp.join(self.nusc_source, filename))
-                is_lidar = 0
-            else:
-                pointcloud = LidarPointCloud.from_file(osp.join(self.nusc_source, filename))
-                is_lidar = 1
-
-            # Transform pointcloud from sensor coordinate frame to ego pose frame
-            rotation_matrix = Quaternion(this_calibrated_sensor['rotation']).rotation_matrix
-            pointcloud.rotate(rotation_matrix)
-            pointcloud.translate(np.array(this_calibrated_sensor['translation']))
-
-            # Rotate velocities to ego pose frame (for radar)
-            if not is_lidar:
-                velocities = pointcloud.points[8:10, :]
-                velocities = np.vstack((velocities, np.zeros(pointcloud.points.shape[1])))
-                pointcloud.points[8:10, :] = np.dot(rotation_matrix, velocities)[:2]
-
-            # transpose for points_in_box and get only first 3 dimensions (coordinates)
-            points_with_metadata = np.transpose(pointcloud.points)
-            points = []
-            for point_data in points_with_metadata:
-                points.append(point_data[:3])
-            points = np.transpose(points)
-            # filter for points in box
-            filter_mask = points_in_box(box=box, points=points)
-            filtered_points = pointcloud.points[:, filter_mask]
-
-            # print(sensor_name)
-            # print(filtered_points)
-
-            # add points from this sensor to the points from the other sensors (in ego pose frame)
-            # points_ego_frame FIELDS: x y z is_lidar intensity vx vy
-            for point in range(len(filtered_points[0])):
-                for i in range(3):
-                    points_ego_frame[i].append(filtered_points[i][point])
-                points_ego_frame[3].append(is_lidar)
-                if is_lidar:
-                    points_ego_frame[4].append(filtered_points[3][point])
-                    points_ego_frame[5].append(0)
-                    points_ego_frame[6].append(0)
+                # Get points (sensor coordinate frame)
+                filename = this_sample_data['filename']
+                if this_sample_data['sensor_modality'] == 'radar':
+                    pointcloud = RadarPointCloud.from_file(osp.join(self.nusc_source, filename))
+                    is_lidar = 0
                 else:
-                    points_ego_frame[4].append(0)
-                    points_ego_frame[5].append(filtered_points[8][point])
-                    points_ego_frame[6].append(filtered_points[9][point])
+                    pointcloud = LidarPointCloud.from_file(osp.join(self.nusc_source, filename))
+                    is_lidar = 1
 
-        label = this_annotation['category_name']
-        return points_ego_frame, label
+                # Transform pointcloud from sensor coordinate frame to ego pose frame
+                rotation_matrix = Quaternion(this_calibrated_sensor['rotation']).rotation_matrix
+                pointcloud.rotate(rotation_matrix)
+                pointcloud.translate(np.array(this_calibrated_sensor['translation']))
+
+                # Rotate velocities to ego pose frame (for radar)
+                if not is_lidar:
+                    velocities = pointcloud.points[8:10, :]
+                    velocities = np.vstack((velocities, np.zeros(pointcloud.points.shape[1])))
+                    pointcloud.points[8:10, :] = np.dot(rotation_matrix, velocities)[:2]
+
+                # transpose for points_in_box and get only first 3 dimensions (coordinates)
+                points_with_metadata = np.transpose(pointcloud.points)
+                points = []
+                for point_data in points_with_metadata:
+                    points.append(point_data[:3])
+                points = np.transpose(points)
+                # filter for points in box
+                filter_mask = points_in_box(box=box, points=points)
+                filtered_points = pointcloud.points[:, filter_mask]
+
+                # print(sensor_name)
+                # print(filtered_points)
+
+                # add points from this sensor to the points from the other sensors (in ego pose frame)
+                # points_ego_frame FIELDS: x y z is_lidar intensity vx vy
+                for point in range(len(filtered_points[0])):
+                    for i in range(3):
+                        points_ego_frame[i].append(filtered_points[i][point])
+                    points_ego_frame[3].append(is_lidar)
+                    if is_lidar:
+                        points_ego_frame[4].append(filtered_points[3][point])
+                        points_ego_frame[5].append(0)
+                        points_ego_frame[6].append(0)
+                    else:
+                        points_ego_frame[4].append(0)
+                        points_ego_frame[5].append(filtered_points[8][point])
+                        points_ego_frame[6].append(filtered_points[9][point])
+
+            label = this_annotation['category_name']
+            return points_ego_frame, label
+        except FileNotFoundError:
+            return None
 
     def process(self):
         torch.save(self.process_set("train"), self.processed_paths[0])
@@ -138,19 +142,26 @@ class NuScenesLoader(Dataset):
 
         for i in range(len(self)):
             print("Currently processing point cloud number {} of {}".format(i, len(self)))
-            points_array, sample_category = self.get_data(i)
+            try:
+                if not Path(osp.join(self.processed_dir, 'data_{}.pt'.format(i))).exists():
+                    points_array, sample_category = self.get_data(i)
 
-            target = categories.index(sample_category)
-            points_array = np.transpose(points_array)
-            data = Data(x=points_array, y=torch.tensor([target]))
+                    target = categories.index(sample_category)
+                    points_array = np.transpose(points_array)
+                    data = Data(x=points_array, y=torch.tensor([target]))
 
-            if self.pre_filter is not None and not self.pre_filter(data):
-                continue
+                    if self.pre_filter is not None and not self.pre_filter(data):
+                        print("File number {} is filtered out (invalid, maybe no points).".format(i))
+                        continue
 
-            if self.pre_transform is not None:
-                data = self.pre_transform(data)
+                    if self.pre_transform is not None:
+                        data = self.pre_transform(data)
 
-            torch.save(data, osp.join(self.processed_dir, 'data_{}.pt'.format(i)))
+                    torch.save(data, osp.join(self.processed_dir, 'data_{}.pt'.format(i)))
+                else:
+                    print("File number {} is skipped (already exists).".format(i))
+            except TypeError:
+                print("File number {} is skipped (data not found).".format(i))
 
     def __len__(self):
         return len(self.dataset.sample_annotation)
